prov/xnet: Return an error if invalid sequence number received
If compiled with --enable-debug, the provider would fail an
assertion if the sequence number received in the header doesn't
match the one expected.

Crashing on an assertion failure is fine for programming errors,
but the assertion failure described above relies on external events:
it may for example trigger because of a buffers corrupted by the
TCP stack or the physical layer.

Instead of an assertion failure, the function returns an error
code, which will cause the EP to be disabled as a consequence.

Signed-off-by: Sylvain Didelot <sdidelot@ddn.com>