core: Convert old fi_info to latest def
fi_dupinfo() for ABI 1.1 and 1.2 was only converting fi_info and none of
the other attr structs inside fi_info. This can lead to current
definition of fi_dupinfo() to read in garbage data from previous struct
definitions. To prevent this dup and convert the all old fi_info and
attrs to the latest definition. Any new fields will be zeroed.

Signed-off-by: Ian Ziemba <ian.ziemba@hpe.com>