prov/efa: Improve efa_rdm_pke definition
This removes the hardcoded padding bytes from the struct definition,
which were ostensibly used as a mechanism to enforce alignment of
wiredata to the 128 byte boundary.

While the static_asserts somewhat protect against changes made to the
members which the struct is composed of, the manual calculation is a
maintenance headache which is prone to developer error, as well as
implementation-defined quirks.

More importantly, this change more clearly indicates the intent to
align to the 128-byte boundary.

This commit also makes the struct size uniform when debug mode is
enabled. This wasn't (currently) necessary, as the 32 bytes of padding
were already sufficient for the 16-byte dlist_entry struct.

This is the C11-compliant version, utilizing _Alignas (stdalign.h)

Signed-off-by: Darryl Abbate <drl@amazon.com>