core: Add restricted dlopen flag to configure options
This commit introduces a new flag, --enable-restricted-dl.
By default, restricted_dl is false, which continues the existing behavior
of fi_getinfo. The current behavior of fi_getinfo causes the library to
probe all paths in LD_LIBRARY_PATH for all possible provider implementations
known by ofi_ordered_provs_init, or in cases where FI_PROVIDER_PATH is defined,
a search along a colon-delimited set of paths for any library named lib*-fi.so.
A side-effect of the non-FI_PROVIDER_PATH option is that a vast majority of the
providers in the ofi_ordered_provs_init list, ordered_prov_names, will not exist
on most systems. Usually a small subset of the providers exist, and the rest will
not be present. A significant number of metadata fetches to the
network filesystem will fail to find the desired library via dlopen incurring
significant penalties during job launch at scale. As an application scales to
larger rank and node counts, the impact by this behavior is expected to grow
linearly, and then exponentially as the network file system metadata servers are
overwhelmed.

This impact of dlopen misses can be largely mitigated with the use of FI_PROVIDER_PATH,
but that eliminates the ability to naturally load DL libraries through dlopen's search
capability unless the admin or workload manager specify possible paths for each known
set of DL provider implementations.

This commit seeks to find a solution in the middle -- dlopen capability and natural
discovery of DL providers without searching for every possible provider and no
prerequisite knowledge of the available DL providers in the workload manager
or module environments.

When enabled, ofi_ordered_provs_init will seed the initial prov_head with
providers that satisfy the following conditions:
- The provider has been built into the library (detected at compile-time
  without --enable-only, or --enable-<provider>[=yes])
- The provider has been built standalone as a DL provider (--enable-<provider>=dl)
- The provider exists in the 'ordered_load_list' in ofi_ordered_provs_init.

Providers that do not meet the above conditions will not be loaded by libfabric
if --enable-restricted-dl is provided to configure.

A similar feature will be required for HMEM provider solutions. Vendors may
need to compile support for GPU libraries, but customers of the vendor may not
use one or more of the supported GPU libraries.

Signed-off-by: James Swaro <james.swaro@hpe.com>