hmem: generalize the use of OFI_HMEM_DATA to non-cuda iface
The use of dev_reg copy is not specific to cuda but can be used by other
iface to fasten the host-to-device and device-to-host memory copy.
Changing the name to a non-cuda specific terminology makes more sense to
reflect the wider use of HMEM_DATA.

The presence of the flag can now be used in generic way to indicate that
hmem_data contains the appropriate handles in order to use
`ofi_hmem_dev_reg_copy_to_hmem` instead of `ofi_hmem_copy_to_hmem`.

Signed-off-by: Thomas Gillis <thoma.gillis@gmail.com>