util: Update util_cq to support err_data
The issue with the current implementation of ofi_cq_write_error() is
that it just reuses the provider provided err_data pointer. This creates
the issue where a provider does not know when it is safe to free the
err_data pointer.

To fix this, if err_data is valid, create a err_data copy internal to
the util_cq code. This internal err_data copy will automatically be
freed in ofi_cq_readerr() once copy into the users buffer.

After a provider calls ofi_cq_write_error(), they can free their
err_data buffer.

Signed-off-by: Ian Ziemba <ian.ziemba@hpe.com>