hmem/cuda: Fallback to CUDA calls when NVML call fails.
The NVML calls to verify nvidia devices are nice-to-have
features that allow a cheaper check and earlier exit when
cuda devices are not present. When NVML call fails, we
shouldn't fail the whole cuda hmem init but should fall
back to the normal cuda api calls.

It is observed that users may use an NVML libraries that
only contain stub functions for build purpose.

Signed-off-by: Shi Jin <sjina@amazon.com>