github/actions: migrate from "hub" to "gh"
GitHub has removed "hub" from their runners since Oct.2, 2023. The
recommendation is to use "gh" instead. One benefits is that "gh"
has a command for merging the PR, with option to rebase. This allows
simpler action script and makes it possible to make "rebase" as the
only merge strategy allowed.

Signed-off-by: Jianxin Xiong <jianxin.xiong@intel.com>