man: Clarify the defition of FI_OPT_MIN_MULTI_RECV
The description of this option was inconsistent with the description of
FI_MULTI_RECV in the fi_msg man page w.r.t. the scenario when the buffer
space available is equal to the size set by the option.

Now make it clear that the buffer is freed when the buffer space available
falls below the set size.

Signed-off-by: Jianxin Xiong <jianxin.xiong@intel.com>