prov/efa: Use bounce buffer for 0 byte writes
The non-inject 0 byte write path assumed that users would pass in
valid buffers, and valid descriptors (memory registrations) for
those buffers. The user should not have to pass in a valid buffer to do
a 0 byte write. Use pre-registered bounce buffers for buffer/MR, to
allow user to pass in whatever they choose.

Signed-off-by: Seth Zegelstein <szegel@amazon.com>