prov/rstream: Fix compiler warning about buffer maybe used uninitialized
The compiler thinks the buffer content is used inside the fi_mr_reg() call
in some way. Since the function accepts the buffer as "const char *" type,
any initialization should be done before the call. When the buffer is not
initialized a warning message is generated.

Even though the fi_mr_reg() call doesn't use the content of the buffer,
replace malloc() with calloc() to eliminate the warning.

Also add an error check for the allocation call.

Signed-off-by: Jianxin Xiong <jianxin.xiong@intel.com>