prov/verbs: Call rdma_resolve_addr() only if node or dst addr is provided
The call to rdma_resolve_addr() in vrb_get_rai_id() would block the
caller for 2 seconds if the remote address is unreachable. Because
this function is executed in the context of fi_getinfo(), we cannot
convert it to a non-blocking call but we can mitigate the issue: we
can assume that the caller is interested in the destination address
only if the node or the dest address (in the hints) is provided.
For some application, the destination address can be exchanged
out-of-bound and does not required to be resolved by fi_getinfo().

Signed-off-by: Sylvain Didelot <sdidelot@ddn.com>