prov/verbs: Always allocate the CM hdr and skip it when route resolved
With the upcoming changes, we will no longer know the sockaddr address
family when fi_connect() is executed. This piece of information will be
known only after when the route is resolved.

The patch updates the code to always allocate the CM header but skip
it when the route is resolved (if needed).

Signed-off-by: Sylvain Didelot <sdidelot@ddn.com>