[v1.15.x] util/memhooks: Fix a bug when calculating mprotect region
In ofi_write_patch, the bound should be the page end for the
end of the address + data_size. And since it's the end of the page,
when calculating the length, it needs to add 1 to make the length
a multiple of page size.

Signed-off-by: Shi Jin <sjina@amazon.com>
(cherry picked from commit faf4ff43ae5ab9c2e0f6f3b8a31b1d6bb142572e)