[v1.19.x] prov/efa: Make runt_size aligned
Currently, txe->runt_size can be non-multiple of the
memory alignment, which caused the following issues:

1. It caused the data size to be copied on the receiver side
not a multiple of memory alignment. This not only caused
non-performant data copy (gdrcopy or local read), but also
breaks the LL128 protocol for send/recv, which requires the
data size to be copied must be a multiple of 128 (the memory
alignment in this case).

2. It caused the single_pkt_entry_data_size variable in
efa_rdm_ope_prepare_to_post_send() to be 0 after doing
the alignment trim.

This patch makes the runt size always aligned before we
decide whether to use runting read protocol. If the
aligned runt size is 0, we won't do runting read.

Also added a series of unit test to validate this change

Signed-off-by: Shi Jin <sjina@amazon.com>
(cherry picked from commit 1604521181433cce426bc5d25788a0e2a89cefb4)