prov/linkx: Make lp_provs a linked list
Make lp_provs a linked list. An array means you have to iterate the entire
array to find what you're looking for. A linked list is better for the type
of search and access needed to the providers.

Also keep a separate field for the shm memory provider for ease of access.

Signed-off-by: Amir Shehata <shehataa@ornl.gov>