prov/efa: Refactor `efa_strerror()`
This reverts efa_strerror() to its original semantics; more closely
matching the ISO C function strerror(). This still effectively allows
the the default error messages to be overridden when the CQ entry has
err_data; it simply removes the superfluous err_data check and NULL
arguments for the common case.

Signed-off-by: Darryl Abbate <drl@amazon.com>