prov/efa: Remove error message assertions from CQ unit tests
Assertions based on the error codes themselves is sufficient. Asserting
the contents of the error message should be outside the scope of the CQ
unit tests. Since the content(s) of the messages don't contain specific
non-static data that need to be correct (i.e. it affects functionality),
these assertions just create unnecessary complexity and maintenance
overhead.

Signed-off-by: Darryl Abbate <drl@amazon.com>