prov/linkx: Optimize lnx_match_common()
lnx_match_common() is called to match a peer entry to a message or a
receive request. This function needs to be as efficient as possible.
However, there are competing requirements. To support Multi-Rail we need
to be able to handle matching against multiple different potential
addresses for a single peer. That adds looping which introduces latencies.
This is an attempt to optimize for the most used scenario, where we link
one provider with SHM.

Signed-off-by: Amir Shehata <shehataa@ornl.gov>