prov/efa: Make handshake response use txe
In order to optimize the fast path send path, we need the handshake
response code to create a txe. Now all sends will have a txe with a
valid peer so there is no need to do the lookup multiple times.

Fixes a bug in the error path to report issue to CQ instead of EQ.

Updates EFA Unit test to make the corresponding changes pass.

Signed-off-by: Seth Zegelstein <szegel@amazon.com>