Fix compiler warning when compiling with ICX
Compiling with icx uncovered an error from -Wenum-conversion. Need
to explicitly cast from CUresult to cudaError_t for ofi_cudaGetErrorName
and ofi_cudaGetErrorString calls.

Even with this fix, there is a risk that the return code is not a valid
cudaError_t enumeration, but the only consequence will be that the
functions return "Unknown Error".

Signed-off-by: Scott Breyer <scott.breyer@intel.com>