prov/util: Make ep_list_lock noop for FI_PROGRESS_CONTROL_UNIFIED
This patch makes ep_list_lock a noop when user has set
FI_PROGRESS_CONTROL_UNIFIED and the threading model is FI_THREAD_DOMAIN
or FI_THREAD_COMPLETION. The util_cntr and util_cq ep_list_locks are used
on the critical path for data transfer, so this change should provide a
performance improvement. The util_av ep_list_lock should not be used on the
performance critical path, but it is changed for simiplicty/consistency for
the reader.

Signed-off-by: Seth Zegelstein <szegel@amazon.com>