src/hmem_ze: dynamically allocate device resources based on number of devices
Instead of hard coding the number of devices and having to increase
it constantly, dynamically allocate arrays of devices based on how
many possible devices there are on the system. To simplify the code,
this groups the device resources (ordinals, indices, cqs, cls) into
a struct so we can allocate everything all at once. The device handles
are still separate because a contiguous array is needed for the device
query call. The dev fds are also separate to prepare for a future fix
to separate the ZE device (used in the ZE library to represent a device)
and the PCI physical device (used in the ioctl IPC implementation).

Note this temporarily adds a place holder ZE_MAX_DEVICES within
the shm provider because it is used to handle IPC device fds.
This will be replaced in another patch as this involves more
invasive changes in the shm provider which are better for a
separate patch.

Signed-off-by: Alexia Ingerson <alexia.ingerson@intel.com>