[v1.19.x] src/hmem_ze: map device id to physical device
The devices returned from zeDeviceGet map to each subdevice (tile),
not physical PCI devices. This means on a system with 2 physical
devices, each with 2 tiles, ZE will return 4 devices total.
This conflicts with the shm IPC implementation which uses device
fds on each physical device.

This commit creates a mapping between the devices returned from
zeDeviceGet and the physical devices they are on and uses that
mapping internally to get the correct device fd for the IPC
implementation.

This is done by looking at the device domain and PCI bus IDs
returned from the device properties (known to be at slots 6 and 8,
respectively) and matching it to the domain and bus IsD in the file
name opened for the device fd, which is in the format
domain:bus:dev.func

Signed-off-by: Alexia Ingerson <alexia.ingerson@intel.com>
(cherry picked from commit dccd54c4f95c2b7c5cc4684768a260630548687a)