prov/shm: adjust the order of smr_region fields.
xpmem is a opt-in feature and not enabled by default.
The previous commit e0906f6f inserts xpmem related fields
in the middle of smr_region struct, which makes the size
of smr_region struct increase from 104 (2 cache lines) to 144
( 3 cache lines). xpmem_cap_self, xpmem_self, and xpmem_peer
are not used in fast path while they make the fields (offsets)
used in the fast path shifted to the next cache line and causes more
cache misses.

This patch moves the xpmem fields to the end of the struct to fix
this issue.

Signed-off-by: Shi Jin <sjina@amazon.com>