prov/efa: Fix ep_close race condition
For an app claiming FI_THREAD_SAFE, a thread can be progressing a cq bound
to the EP at the same time another thread calls fi_close() on the ep.
To avoid this race condition, we need to close the util_ep before we
free any efa provider EP resources. Freeing the util_ep will grab the
ep_list_lock which will guarantee there are no other threads are actively
progressing the CQ, and then remove the ep from the cq's ep_list.

Signed-off-by: Seth Zegelstein <szegel@amazon.com>