contrib/intel/jenkins: Add missing test names to shmem summary
SHMEM is missing PRK and ISx test names from the summary.
This patch adds the capability to find the test names and add
them to the passed and failed test lists for verbose reporting.
Tests will have the type of SHMEM test added as a prefix to them.

For Example:
            UH test_1
            PRK test_1
            ISx test_1

Signed-off-by: Juee Himalbhai Desai <juee.himalbhai.desai@intel.com>