prov/efa: Handle rx pkts error without ope
When getting the error with IBV_WC_RECV op
code, it is possible that packet entry is not associated
with an ope yet, if the packet is posted by
efa_rdm_ep_bulk_post_internal_rx_pkts().

For such error, we can only warn and write eq
error as it is an EFA internal error and not
associated with application's rx operation.

Signed-off-by: Shi Jin <sjina@amazon.com>