prov/shm: Lock double deletion check
This change solves this coverity issue:
Checking if an AV (address vector) map has already deleted a peer, or if the
peer no longer exists in the AV map must be locked. This is because one
thread can check if a peer exists and get a "false positive" while another
thread is actively destroying that peer. This case would result in a double
destruction which we need to protect against.

Signed-off-by: Zach Dworkin <zachary.dworkin@intel.com>