prov/cxi: Add experimental truncation as a success completions
Add the ability for truncation to be reported as success completions,
passing an additional flag FI_TRUNC and the truncated length.

This is experimental and is only supported by the client/server
protocol. Ultimately a new FI_TRUNC_OK mode bit is proposed
to make this behavior always turned on, but for early testing
set FI_CXI_TRUNC_OK=1.

If a receive is truncated and completions are requested, the
truncated length will be returned in a successful completion
with the addition flag FI_CXI_TRUNC. For a send, it means
that the length will be reported in a send completion, and
if the target reported a truncation, then the truncated length
and FI_CXI_TRUNC flag will be reported in the completion.

NETCASSINI-5652
NETCASSINI-6125

Signed-off-by: Steve Welch <welch@hpe.com>
(cherry picked from commit d6892bbb64348c567216a2656177780baa192f50)