prov/hook: Proposal for updating the hook provider
As far as I could tell the hook provider dumps all the trace to the
console, which is a lot to look at. This is a proposal to store the
tracing information internally and then dump it when shutting down.

This patch makes a change to how the send/recv operation tracing is done.

This example code stores a record in a database. The record has a key
consistent of addr/data len/operation. This way we can keep track of the
number of operations of a specific len per address. I then dump it to a
csv file, which is a lot easier to parse. Typically we don't need to see
each operation on the console. A summary statistics of the job would be
more useful (IMHO)

This is just a proposal to get some feedback.

Signed-off-by: Amir Shehata <shehataa@ornl.gov>