prov/shm: manually align 8 byte fields in memory region
Move all of the 8 byte fields together to eliminate extra padding.
Previously, the last 8 byte field was after a 4 byte field, forcing the
compiler to add extra padding before the 4 byte field and after the third
8 byte field. This increased the size of the memory region and pushed some
fields to a new cache line. Consolidating them eliminates those extra bytes
and allows for better caching performance.
This also adds a reserved value to force alignment of the next 32 bit field
though the compiler should automatically add that memory anyway.

Altering the order of the fields in the shared memory region requires a
version increment.

Signed-off-by: Alexia Ingerson <alexia.ingerson@intel.com>