hmem/cuda: Get multiple attributes at once in cuda_is_addr_valid
The cuda_is_addr_valid() function makes 2 separate calls to
cuPointerGetAttribute() to get the memory type and device. If the
caller also wants to know if the cuda buffer is "managed"
(i.e. unified memory), a 3rd subsequent call to cuPointerGetAttribute()
is necessary (outside of cuda_is_addr_valid()) to retrieve that.

This change gets all three attributes in a single call to
cuPointerGetAttributes(). If the address in question points
to cuda memory, and that memory is managed/unified, the
resulting flags will not have FI_HMEM_DEVICE_ONLY set.

Signed-off-by: Ben Lynam <Ben.Lynam@cornelisnetworks.com>