prov/sm2: fix uninitialized iov warning
Coverity is complaining that iov, result_iov, and compare_iov
are not getting initialized before getting used later.
However, they should always get initialized in their respective op
The path it shows is for an incorrect atomic op.
This adds an error case to prevent coverity from thinking
it can continue with an incorrect atomic op

Signed-off-by: Alexia Ingerson <alexia.ingerson@intel.com>