prov/efa: Delegate iovec offset calc to utility function
This should correctly point the IOV base pointer to the start of the
data buffer in prefix mode. This accommodates non-contiguous iovec
entries where iov[0] is the prefix and iov[1] points the data buffer, as
well as the previously-supported behavior of a single iovec entry where
the buffer has ample space for the prefix and user data.

Signed-off-by: Darryl Abbate <drl@amazon.com>