prov/usnic: fix CID 380119
Cast time(NULL) down to 32 bits.  For this specific case, the time is
used as a sequence number.  It is definitely expected that the query
will complete within 2^31 additional querys down to libnl (if it
doesn't, we have larger issues), so even if sizeof(time_t)==8, it's
ok if we truncate it down to the lower 32 bits.

Signed-off-by: Jeff Squyres <jsquyres@cisco.com>