prov/efa: Find offset before posting user recv buffers
This finds the offset of the user data before posting the receive
buffers via ibv_post_recv(). This accommodates cases where some user
data exists in the space allocated for the prefix in as well as the
payload itself.

This also re-adds the second SGE in the ibv_recv_wr wrapper (struct
efa_recv_wr).

Signed-off-by: Darryl Abbate <drl@amazon.com>
Co-authored-by: Shi Jin <sjina@amazon.com>