prov/efa: Do not assert tx and rx msg order must equal
The tx_attr at the sender should align with the rx_attr at the receiver.
But that's an application level responsibility to handle.
The tx_attr and rx_attr of a single endpoint are likely to be consistent,
but that's not required. Efa doesn't have to generate an error
when an inconsitency is detected, but can generate a warning for it.

Also added unit tests.

Signed-off-by: Shi Jin <sjina@amazon.com>