srx: Add memory registration cb to fi_ops_srx_peer
Add a memory registration callback to the fi_ops_srx_peer. This allows
core providers to expose a memory registration callback which the parent
or peer provider can use to register memory on the receive path.

For example the CXI provider registers memory with the NIC on the receive
path. When using the peer infrastructure this can not happen because we
do not know which provider will perform the receive operation. But if
the source NID is specified then we can know and therefore we can
perform the receive buffer registration at the top of the receive path.

Signed-off-by: Amir Shehata <shehataa@ornl.gov>