prov/linkx: Update the way we do SRX
Upstream has a different method of registering SRX. There is a limitation
where the SRX is only returned back in the upcall in get_tag/get_msg. But
that prevents the parent provider of doing anything else with the
peer callbacks. This presents a problem because we added a callback to
register memory on the receive path.

This patch updates the LINKx code

Signed-off-by: Amir Shehata <shehataa@ornl.gov>