prov/efa: trigger handshake with handshake pkt directly.
Currently, we trigger a handshake by making sender sends
a REQ pkt and receiver sends back the handshake pkt as an ACK.
This uni-directional handshake is not necessary and involves 4
ctrl pkts between a pair of peers if both of them want to sync
with the other.

This patch improves this protocol by allowing sender to trigger
a handshake by sending a handshake pkt directly. Then there are
at most 2 ctrl pkts between a peer to make a bi-directional handshake.

Signed-off-by: Shi Jin <sjina@amazon.com>