rdma/fi_domain: Add page size MR attr field
Page size allows applications to optionally notify the provider the page
size to be used for an MR allocation. Typically, providers can select
the optimal page size. In cases where a VA range has zero pages backing
it, the provider may not know the optimal page size during registration.
Rather than always use a less efficient page size, allow apps to specify
the page size to be used. If page size is zero, provider will select the
page size.

If non-zero, page size must be page size support by OS. If a specific
page size is specified for a memory region during creation, all pages
later associated with the region must be of the given size. Attaching a
memory page of a different size to a region may result in failed
transfers to or from the region.

Signed-off-by: Ian Ziemba <ian.ziemba@hpe.com>