prov/efa: Fix the rx pkt post procedure.
Currently, we only repost rx pkts when there is
completion of IBV_WC_RECV and IBV_WC_RDMA_WRITE_IMM.
This is wrong because the rx pkt can be resolved
after being held for local read, which is a send
or read completion. This patch fixes this issue
by just checking the efa_rx_pkts_to_post counter
in ep instead of checking the IBV op codes, which
is safe and stable.

Signed-off-by: Shi Jin <sjina@amazon.com>