prov/efa: Introduce extra feature/request for user recv qp
This bit is interpreted as a feature on the sender side,
and a request on the receiver side. When zero copy mode
is turned on, sender can choose to send data to receiver's
user recv QP without hdr to optimize performance.
For receiver, when zero-copy mode is enabled, it can only
get pkts delivered to the user recv QP.

Signed-off-by: Shi Jin <sjina@amazon.com>