prov/verbs: Optimize search for device max inline size
Let the user set the FI_VERBS_INLINE_SIZE environment variable to
decide the maximum size of an inline transfer. The getinfo call will
check if the device is capable of this size. If it is capable it will
use it. If it not capable then the size is too big and we can goto the
backwards search to find the largest inline size that is smaller than
the FI_VERBS_INLINE_SIZE variable.

If a user wants the maximum inject size that a device can do then they
should set this FI_VERBS_INLINE_SIZE variable to 0 so that the device
can find its largest inject size.

Signed-off-by: Zach Dworkin <zachary.dworkin@intel.com>