fabtests: Syncronize on Initialization fabtests/functional: Add manual init sync to fi_rdm_multiclient and fi_rdm
Some providers (verbs ud) might require the server to be fully initialized
before the client process calls getinfo with the server address.
This causes a No Data Available error due to the fi_info call failing during
initialization by not being able to find the name on the server.
This is seen most often in cases where a socket (usually oob [out of band]) is
initialized before getinfo in the ft_init_fabric sequence. Adding a sync only
if an oob socket has been initialized to order the initialization correctly will
prevent this from happening.

The syncronization is for client to start getinfo only after the server is done
initializing everything.

fi_rdm_multiclient test needs its client startup to have a manual sync since it
does not follow the normal codepath of going through ft_init_fabric like the server
and other tests do. This sync will only happen on the first client that connects
because it is only necessary to give the server enough time to spin up all the
resources. It is not necessary for future clients because the server has already
started all of those resources.

fi_rdm needs the syncronization added after accepting a new client because the
client will be waiting for a sock_send and the server only does it once on its
initialization. Adding a ft_sock_sync will force the server to do it for each
new client.

Signed-off-by: Zach Dworkin <zachary.dworkin@intel.com>